backend/requirements.txt
# FastAPI and server
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.9

# Pydantic and settings
pydantic>=2.6.0
pydantic-settings>=2.1.0
email-validator>=2.1.0

# Security
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4

# Supabase
supabase>=2.27.0
websockets>=11,<16

# Rate limiting
slowapi>=0.1.9

# Logging
python-json-logger>=2.0.7
